refactor: create lotto.result pkg
