+++ src/main/java/lotto/result/LOTTO_RESULT.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.result;
 
 import lotto.number.LottoNumbers;
