+++ src/main/java/lotto/result/LottoResultMatcher.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.result;
 
 interface LottoResultMatcher {
 	boolean isMatched(int matchedCount);
