+++ src/test/java/lotto/LottoGameTest.java
@@ -1,6 +1,7 @@
 package lotto;
 
 import lotto.number.LottoNumbers;
+import lotto.result.LottoStatistics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
